requirements.txt
asgiref==3.7.2
Django==4.2.4
django-cors-headers==4.2.0
django-sendgrid-v5==1.2.2
djangorestframework==3.14.0
python-http-client==3.3.7
pytz==2023.3
sendgrid==6.10.0
sqlparse==0.4.4
starkbank-ecdsa==2.2.0
typing_extensions==4.7.1
